+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Joyride
 
 ## [Unreleased]
 
+- [Add ”Evaluate Selection” command](https://github.com/BetterThanTomorrow/joyride/issues/106)
+
 ## [0.0.22] - 2022-11-18
 
 - [Make nREPL host address configurable](https://github.com/BetterThanTomorrow/joyride/issues/102)